Add avl tests, change module names -add tests for all public methods in avl tree -renamed packages -separated general and specific avl tests (created 2 classes)
